torrentbot/requirements.txt
python-telegram-bot==21.6
anthropic==0.40.0
pyyaml==6.0.2
aiohttp==3.11.0
aiosqlite==0.20.0
python-dotenv>=1.1.0
